ExecutionService の checkpoint 永続化を ExecutionCheckpointService に分離し、Unload ゲートと世代フェンスの責務を独立させる。
