+++ core/application/Statevia.Core.Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -29,6 +29,7 @@ public static class ApplicationServiceCollectionExtensions
         services.AddScoped<ExecutionProjectionOrchestrator>();
         services.AddScoped<ExecutionLifecycleCommandService>();
         services.AddScoped<ExecutionWaitEventService>();
+        services.AddScoped<ExecutionCheckpointService>();
         services.AddScoped<IExecutionService, ExecutionService>();
         services.AddScoped<IEventIngressService, EventIngressService>();
         services.AddScoped<IForkChildExecutionCoordinator, ForkChildExecutionCoordinator>();
